int chamge kore float

+++ Calculator.py
@@ -2,7 +2,7 @@ print("1 - Add")
 print("2 - Subtract")
 print("3 - Multiply")
 print("4 - Divide")
-option = int(input("Choose an operation: "))
+option = float(input("Choose an operation: "))
 
 if(option in [1,2,3,4]):
     number1 = float(input("Enter the first number:"))
